requirements.txt
selenium
pandas
python-dotenv
pytest